LPS-57084 Move
